Put message rendering views in their own folder
