+++ Swiftcord/Views/Message/MessageRenderViews/MessageView.swift
@@ -35,85 +35,6 @@ struct NonUserBadge: View {
 	}
 }
 
-/// Action messages: e.g. member leave, join etc.
-struct ActionMessageView: View {
-	let message: Message
-
-	private struct ActionMessageData {
-		let message: LocalizedStringKey
-		let icon: String
-		let color: Color
-	}
-
-	// Trust, this might seem messy but it's more scalable than putting the
-	// views themselves in a huge if tree
-	private var data: ActionMessageData {
-		switch message.type {
-		case .guildMemberJoin:
-			return ActionMessageData(
-				message: "**\(message.author.username)** joined this server.",
-				icon: "arrow.right",
-				color: .green
-			)
-		case .recipientAdd:
-			return ActionMessageData(
-				message: "**\(message.author.username)** added **\(message.mentions[0].username)** to the group.",
-				icon: "arrow.right",
-				color: .green
-			)
-		case .recipientRemove:
-			return ActionMessageData(
-				message: "**\(message.author.username)** left the group.",
-				icon: "arrow.left",
-				color: .red
-			)
-		case .userPremiumGuildSub:
-			return ActionMessageData(
-				message: "**\(message.author.username)** just boosted the server!",
-				icon: "rhombus.fill",
-				color: .purple
-			)
-		case .userPremiumGuildSubTier1:
-			return ActionMessageData(
-				message: "**\(message.author.username)** just boosted the server! This server has achieved **Level 1!**",
-				icon: "rhombus.fill",
-				color: .purple
-			)
-		case .userPremiumGuildSubTier2:
-			return ActionMessageData(
-				message: "**\(message.author.username)** just boosted the server! This server has achieved **Level 2!**",
-				icon: "rhombus.fill",
-				color: .purple
-			)
-		case .userPremiumGuildSubTier3:
-			return ActionMessageData(
-				message: "**\(message.author.username)** just boosted the server! This server has achieved **Level 3!**",
-				icon: "rhombus.fill",
-				color: .purple
-			)
-		default:
-			return ActionMessageData(
-				message: "Oops, rendering `\(String(describing: message.type))` messages aren't supported yet :(",
-				icon: "questionmark",
-				color: .gray
-			)
-		}
-	}
-
-	var body: some View {
-		Image(systemName: data.icon)
-			.foregroundColor(data.color)
-			.font(.system(size: 16))
-			.padding([.leading, .trailing], 12)
-		Group {
-			Text(data.message).font(.system(size: 14))
-			+ Text(" ").font(.system(size: 14))
-			+ Text(DateFormatter.messageDateFormatter.string(from: message.timestamp))
-				.font(.system(size: 12))
-		}.opacity(0.75)
-	}
-}
-
 struct MessageView: View {
     let message: Message
     let shrunk: Bool
@@ -131,20 +52,24 @@ struct MessageView: View {
     @EnvironmentObject var serverCtx: ServerContext
 	@EnvironmentObject var restAPI: DiscordREST
 
+	// Messages that can be rendered as "default" messages
+	static let defaultTypes: [MessageType] = [.defaultMsg, .reply]
+
     var body: some View {
         VStack(alignment: .leading, spacing: 2) {
             // This message is a reply!
             if message.type == .reply {
-				ReferenceMessageView(referencedMsg: message.referenced_message)
-					.onTapGesture { if let referencedID = message.referenced_message?.id {
+				ReferenceMessageView(referencedMsg: message.referenced_message).onTapGesture {
+					if let referencedID = message.referenced_message?.id {
 						onQuoteClick(referencedID)
-					}}
+					}
+				}
             }
             HStack(
-                alignment: message.type == .guildMemberJoin || message.type == .userPremiumGuildSub ? .center : .top,
+				alignment: MessageView.defaultTypes.contains(message.type) ? .top : .center,
                 spacing: 16
             ) {
-                if message.type == .reply || message.type == .defaultMsg {
+                if MessageView.defaultTypes.contains(message.type) {
                     if !shrunk {
                         UserAvatarView(user: message.author, guildID: serverCtx.guild!.id, webhookID: message.webhook_id, clickDisabled: false)
                     } else {
@@ -178,46 +103,7 @@ struct MessageView: View {
 								.opacity(0.5)
                             }
                         }
-                        // For including additional message components
-                        VStack(alignment: .leading, spacing: 4) {
-                            if !message.content.isEmpty {
-                                // Guard doesn't work in a view :(((
-                                /*if let msg = attributedMessage(content: message.content) {
-                                    Text(msg)
-                                        .font(.system(size: 15))
-                                        .textSelection(.enabled)
-                                 // fix this poor implementation later
-                                }*/
-								let msg = message.content.containsOnlyEmojiAndSpaces
-									? message.content.replacingOccurrences(of: " ", with: " ")
-									: message.content
-                                Group {
-									Text(markdown: msg)
-									.font(.system(
-										size: message.content.containsOnlyEmojiAndSpaces ? 48 : 15
-									)) +
-                                    Text(
-										message.edited_timestamp != nil && shrunk
-                                         ? "message.edited.shrunk"
-                                         : ""
-									)
-									.font(.system(size: 8))
-									.italic()
-									.foregroundColor(Color(NSColor.textColor).opacity(0.4))
-                                }.textSelection(.enabled)
-                            }
-                            if let stickerItems = message.sticker_items {
-                                ForEach(stickerItems) { sticker in
-                                    StickerView(sticker: sticker)
-                                }
-                            }
-                            ForEach(message.attachments) { attachment in
-                                AttachmentView(attachment: attachment)
-                            }
-                            ForEach(message.embeds) { embed in
-                                EmbedView(embed: embed)
-                            }
-                        }
+						DefaultMessageView(message: message, shrunk: shrunk)
                     }
 				} else {
 					ActionMessageView(message: message)
